ISS790: Refactored lib/object to use .d.ts file rather than embedded typescript
